+++ CHANGELOG.md
@@ -19,3 +19,4 @@
 #### Additions
 
 - Created the literal bare minimum Minecraft world (to minimise world size).
+- Created a script to format some files.